backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
httpx==0.28.1
pydantic==2.11.7
mypy==1.17.1
python-dotenv==1.1.1
python-jose==3.5.0
asyncpg==0.30.0
bcrypt==4.3.0
python-multipart==0.0.20
